fix(git): get rid of erroneous strike-through
Apparently GitHub's markdown syntax interpreted some of the `~` symbols
I used in Git Basics as if they were supposed to represent a
strike-through of some of the text in there. Needless to say, that is
not what I meant.

This commit *should* fix the issue, although I could never get my Common
Mark renderer to replicate the error in the first place, so I honestly
have no idea. Here goes nothing!

Signed-off-by: Buckley Ross <brr322@lehigh.edu>
Change-Id: Id5c4f76eb890175c4eb3563a6d8a41215d5579bf